+++ Tests/Integration/MSFT_ADGroup.Integration.Tests.ps1
@@ -76,7 +76,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group1_Name
                 $resourceCurrentState.GroupScope | Should -Be 'Global'
                 $resourceCurrentState.Category | Should -Be 'Security'
-                $resourceCurrentState.Path | Should -Be 'CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -BeNullOrEmpty
                 $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -137,7 +137,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group2_Name
                 $resourceCurrentState.GroupScope | Should -Be $ConfigurationData.AllNodes.Group2_Scope
                 $resourceCurrentState.Category | Should -Be 'Security'
-                $resourceCurrentState.Path | Should -Be 'CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -BeNullOrEmpty
                 $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -198,7 +198,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group3_Name
                 $resourceCurrentState.GroupScope | Should -Be $ConfigurationData.AllNodes.Group3_Scope
                 $resourceCurrentState.Category | Should -Be 'Security'
-                $resourceCurrentState.Path | Should -Be 'CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -BeNullOrEmpty
                 $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -259,7 +259,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group3_Name
                 $resourceCurrentState.GroupScope | Should -Be $ConfigurationData.AllNodes.Group3_Scope
                 $resourceCurrentState.Category | Should -Be 'Distribution'
-                $resourceCurrentState.Path | Should -Be 'CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -BeNullOrEmpty
                 $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -320,7 +320,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group4_Name
                 $resourceCurrentState.GroupScope | Should -Be $ConfigurationData.AllNodes.Group4_Scope
                 $resourceCurrentState.Category | Should -Be 'Security'
-                $resourceCurrentState.Path | Should -Be 'CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -BeNullOrEmpty
                 $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -381,7 +381,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group4_Name
                 $resourceCurrentState.GroupScope | Should -Be 'Global'
                 $resourceCurrentState.Category | Should -Be 'Security'
-                $resourceCurrentState.Path | Should -Be 'CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -BeNullOrEmpty
                 $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -503,7 +503,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group1_Name
                 $resourceCurrentState.GroupScope | Should -Be 'Global'
                 $resourceCurrentState.Category | Should -Be 'Security'
-                $resourceCurrentState.Path | Should -Be 'CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -BeNullOrEmpty
                 $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -564,7 +564,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group1_Name
                 $resourceCurrentState.GroupScope | Should -Be 'Global'
                 $resourceCurrentState.Category | Should -Be 'Security'
-                $resourceCurrentState.Path | Should -Be 'CN=Computers,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Computers,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -Be 'A DSC description'
                 $resourceCurrentState.DisplayName | Should -Be 'DSC Group 1'
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -575,7 +575,7 @@ try
                 $resourceCurrentState.MembersToInclude | Should -BeNullOrEmpty
                 $resourceCurrentState.MembersToExclude | Should -BeNullOrEmpty
                 $resourceCurrentState.MembershipAttribute | Should -Be 'SamAccountName'
-                $resourceCurrentState.ManagedBy | Should -Be 'CN=Administrator,CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.ManagedBy | Should -Be ('CN=Administrator,CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Notes | Should -Be 'Notes for this group'
                 $resourceCurrentState.RestoreFromRecycleBin | Should -BeNullOrEmpty
             }
@@ -627,7 +627,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group5_Name
                 $resourceCurrentState.GroupScope | Should -Be $ConfigurationData.AllNodes.Group5_Scope
                 $resourceCurrentState.Category | Should -Be $ConfigurationData.AllNodes.Group5_Category
-                $resourceCurrentState.Path | Should -Be 'CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -BeNullOrEmpty
                 $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -689,7 +689,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group5_Name
                 $resourceCurrentState.GroupScope | Should -Be $ConfigurationData.AllNodes.Group5_Scope
                 $resourceCurrentState.Category | Should -Be $ConfigurationData.AllNodes.Group5_Category
-                $resourceCurrentState.Path | Should -Be 'CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -BeNullOrEmpty
                 $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -753,7 +753,7 @@ try
                 $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group5_Name
                 $resourceCurrentState.GroupScope | Should -Be $ConfigurationData.AllNodes.Group5_Scope
                 $resourceCurrentState.Category | Should -Be $ConfigurationData.AllNodes.Group5_Category
-                $resourceCurrentState.Path | Should -Be 'CN=Users,DC=contoso,DC=com'
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
                 $resourceCurrentState.Description | Should -BeNullOrEmpty
                 $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
                 $resourceCurrentState.Credential | Should -BeNullOrEmpty
@@ -761,6 +761,67 @@ try
                 $resourceCurrentState.Members | Should -HaveCount 2
                 $resourceCurrentState.Members | Should -Contain 'Administrator'
                 $resourceCurrentState.Members | Should -Contain 'Guest'
+                $resourceCurrentState.MembersToInclude | Should -BeNullOrEmpty
+                $resourceCurrentState.MembersToExclude | Should -BeNullOrEmpty
+                $resourceCurrentState.MembershipAttribute | Should -Be 'SamAccountName'
+                $resourceCurrentState.ManagedBy | Should -BeNullOrEmpty
+                $resourceCurrentState.Notes | Should -BeNullOrEmpty
+                $resourceCurrentState.RestoreFromRecycleBin | Should -BeNullOrEmpty
+            }
+
+            It 'Should return $true when Test-DscConfiguration is run' {
+                Test-DscConfiguration -Verbose | Should -Be 'True'
+            }
+        }
+
+        $configurationName = "$($script:dscResourceName)_ClearMembersGroup5_Config"
+
+        Context ('When using configuration {0}' -f $configurationName) {
+            It 'Should compile and apply the MOF without throwing' {
+                {
+                    $configurationParameters = @{
+                        OutputPath        = $TestDrive
+                        # The variable $ConfigurationData was dot-sourced above.
+                        ConfigurationData = $ConfigurationData
+                    }
+
+                    & $configurationName @configurationParameters
+
+                    $startDscConfigurationParameters = @{
+                        Path         = $TestDrive
+                        ComputerName = 'localhost'
+                        Wait         = $true
+                        Verbose      = $true
+                        Force        = $true
+                        ErrorAction  = 'Stop'
+                    }
+
+                    Start-DscConfiguration @startDscConfigurationParameters
+                } | Should -Not -Throw
+            }
+
+            It 'Should be able to call Get-DscConfiguration without throwing' {
+                {
+                    $script:currentConfiguration = Get-DscConfiguration -Verbose -ErrorAction Stop
+                } | Should -Not -Throw
+            }
+
+            It 'Should have set the resource and all the parameters should match' {
+                $resourceCurrentState = $script:currentConfiguration | Where-Object -FilterScript {
+                    $_.ConfigurationName -eq $configurationName `
+                        -and $_.ResourceId -eq $resourceId
+                }
+
+                $resourceCurrentState.Ensure | Should -Be 'Present'
+                $resourceCurrentState.GroupName | Should -Be $ConfigurationData.AllNodes.Group5_Name
+                $resourceCurrentState.GroupScope | Should -Be $ConfigurationData.AllNodes.Group5_Scope
+                $resourceCurrentState.Category | Should -Be $ConfigurationData.AllNodes.Group5_Category
+                $resourceCurrentState.Path | Should -Be ('CN=Users,{0}' -f $ConfigurationData.AllNodes.DomainDistinguishedName)
+                $resourceCurrentState.Description | Should -BeNullOrEmpty
+                $resourceCurrentState.DisplayName | Should -BeNullOrEmpty
+                $resourceCurrentState.Credential | Should -BeNullOrEmpty
+                $resourceCurrentState.DomainController | Should -BeNullOrEmpty
+                $resourceCurrentState.Members | Should -BeNullOrEmpty
                 $resourceCurrentState.MembersToInclude | Should -BeNullOrEmpty
                 $resourceCurrentState.MembersToExclude | Should -BeNullOrEmpty
                 $resourceCurrentState.MembershipAttribute | Should -Be 'SamAccountName'